+++ storybook/src/Overlays/Popovers/EdgeOverflow.tsx
@@ -49,17 +49,9 @@ export const EdgeOverflow: FC<Props> = ({
         </PopoverContent>
       }
     >
-      {(onClick, ref, className) => (
-        <ButtonOutline
-          iconAfter="ArrowDown"
-          m="xxlarge"
-          className={className}
-          onClick={onClick}
-          ref={ref}
-        >
-          {children}
-        </ButtonOutline>
-      )}
+      <ButtonOutline iconAfter="ArrowDown" m="xxlarge">
+        {children}
+      </ButtonOutline>
     </Popover>
   </Box>
 )